pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "clide"
authors = [
    {name = "Daniele Affinita", email = "daniele.affinita@sony.com"}
]
description = "Foundation Edge Distillation Architecture"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "wheel",
    "torch",
    "numpy",
    "transformers",
    "overrides",
    "pillow",
    "einops",
    "timm",
    "pymongo", 
    "av",
    "humanfriendly",
    "ultralytics",
    "cython",
    "omegaconf",
    "hydra-core",
    "torchvision",
    "paramiko",
    "scp",
    "model-compression-toolkit>=2.2.0",
    "sony_custom_layers"
]